re-implement the migration logic in the config utilities
