+++ src/vs/editor/browser/config/migrateOptions.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { asBoolean } from '../../../platform/prompts/common/config.js';
 import { IEditorOptions } from '../../common/config/editorOptions.js';
 
 export interface ISettingsReader {
@@ -108,49 +107,59 @@ registerEditorSettingMigration('autoClosingBrackets', (value, read, write) => {
 	}
 });
 
-registerEditorSettingMigration('chat.promptFiles', (value, _read, write) => {
-	if ((value === undefined) || (value === null)) {
-		return;
-	}
 
-	if (Array.isArray(value)) {
-		write('chat.promptFiles', true);
+// TODO: @legomushroom - remove?
+// registerEditorSettingMigration('chat', (_value, read, write) => {
+// 	const value = read('chat.promptFiles');
+// 	const value2 = read('promptFiles');
 
-		const locationsValue: Record<string, boolean> = {};
-		for (const filePath of value) {
-			if (typeof filePath !== 'string') {
-				continue;
-			}
-			const trimmedValue = filePath.trim();
-			if (!trimmedValue) {
-				continue;
-			}
+// 	if ((value === undefined) || (value === null)) {
+// 		return;
+// 	}
 
-			locationsValue[trimmedValue] = true;
-		}
+// 	// // TODO: @legomushroom
+// 	// if (2 + 2 === 4) {
+// 	// 	return;
+// 	// }
 
-		write('chat.promptFilesLocations', locationsValue);
-		return;
-	}
+// 	if (Array.isArray(value)) {
+// 		write('chat.promptFiles', true);
 
-	if (typeof value === 'object') {
-		write('chat.promptFiles', true);
-		write('chat.promptFilesLocations', value);
+// 		const locationsValue: Record<string, boolean> = {};
+// 		for (const filePath of value) {
+// 			if (typeof filePath !== 'string') {
+// 				continue;
+// 			}
+// 			const trimmedValue = filePath.trim();
+// 			if (!trimmedValue) {
+// 				continue;
+// 			}
 
-		return;
-	}
+// 			locationsValue[trimmedValue] = true;
+// 		}
 
-	if (typeof value === 'string') {
-		const booleanValue = asBoolean(value);
-		if (booleanValue !== undefined) {
-			return;
-		}
+// 		write('chat.promptFilesLocations', locationsValue);
+// 		return;
+// 	}
 
-		write('chat.promptFiles', true);
-		write('chat.promptFilesLocations', { [value]: true });
-		return;
-	}
-});
+// 	if (typeof value === 'object') {
+// 		write('chat.promptFiles', true);
+// 		write('chat.promptFilesLocations', value);
+
+// 		return;
+// 	}
+
+// 	if (typeof value === 'string') {
+// 		const booleanValue = asBoolean(value);
+// 		if (booleanValue !== undefined) {
+// 			return;
+// 		}
+
+// 		write('chat.promptFiles', true);
+// 		write('chat.promptFilesLocations', { [value]: true });
+// 		return;
+// 	}
+// });
 
 registerEditorSettingMigration('renderIndentGuides', (value, read, write) => {
 	if (typeof value !== 'undefined') {
